Workaround for unavailable Sign In button in Tab Tray
Workaround for strange behavior of UIKIt w.r.t. its implementation of
UIAccessibility. UIButtons created with .buttonWithType and set as
titleView of UIBarButtonItem which is then added to a UINavigationBar
retain their initial frame height of 0 from the accessibility point of
view.

So set the value initially to some reasonable value.