Bug 1151387 - Part 2: Hide toolbar and web view in home panel
If this is not done, accessibility can still access the toolbar (back,
forward, bookmark, share button) and web view (i.e. contents of the web
page) while not showing them when showing the home panel view. Visual
interface gets by by setting such constraints that stretch the home
panel view to cover both the web view and the toolbar (if displayed
inline). But that does not translate to hiding the elements to
accessibility.

An alternative could be to set `accessibilityElementsHidden` instead of
`hidden`. But the elements should be hidden also visually, so if the
calls in this commit are in the right place (based on my testing they
are), setting `hidden` is correct both visually and accessibility-wise.