Bug 1162174 – Allow Third Party Password Manager Extensions.
This commit moves:
 * the share extension of the BrowserViewController.
 * add OnePasswordExtension to Carthage
 * users OnePasswordExtension to detect and fill password manager extensions.

Tested on OnePassword, LastPass and pwSafe.

Current support of LastPass lags that of OnePassword, though does seem stable, but relies on Copy/Paste.

LastPass should support `browser-fill-action` of the OnePasswordExtension API in order for this to work for without using the clipboard.

The version of one-password-extension is 1.6.4.