Bug 1214794 - ensure that when we are performing a sync whilst deleting existing top sites that we add the correct number of cells
This leads to an immediate problem that I can see - yes nothing crashes, but top sites are then out of order. Newly synced top sites that have a greater weight than existing top sites are displayed _after_ the exiting sites as they are simply added onto the end. Perhaps it would be better in this case to delete _all_ existing top sites rows and repopulate entire collection view with the new data?