Reposition the share sheet after a rotation on iPad.
iOS does not automatically handle reposition of popovers after a
rotation (except when presented from a UIBarButtonItem).

This patch fixes the issue by retaining the displayed popover controller
and a configuration block. These will be used at rotation events to
dismiss the controller then re-present it after the rotation is finished
and after calling the `updateDisplayedPopoverProperties` block to reset
the popover anchor point.