Bug 1228611 - Part 3: use sqlite3_close_v2 to allow database closing to happen later if not all prepared statements have been finalized.
We were using sqlite3_close, which simply returns failure and leaves the
handle open. In that case we'd leak the handle and subsequently open
another connection.

This way we might not close the handle immediately, but we will
eventually, after the last prepared statement has exited scope and been
finalized through its deinit method.