Bug 1277572 – Change new tab behaviour based on user setting (#1916)
* Bug 1277572 – Don't show the home panels if new tab is either a blank or homepage.

This would otherwise mean that there is zero difference between a blank page and the defaults.

* Bug 1277572 – Detect that this is a new tab and load the correct URL into the tab.

* Bug 1277572 – Remove concept of default homepage for opening new tabs.

This replaces the notion of a single unified new tab page set at start up..

* Bug 1277572 – Added defaults for China.

Using literals, because I don't like copying PrefsDefaults and PrefsKeys around, and we can't have a dependency from Profile to Client.

* Bug 1277572 – Update feature flags now there is real functionality available.

* Bug 1277572 – Construct localhostURL from rawValue and remove other calculations of the URL.

* Bug 1277572 – Moved string literals into existing PrefsKeys and PrefsDefaults.