Bug 1228087 - Build a custom UIView to use as a notification when the user enters the app with a url on their clipboard (#2447) r=farhan
* Bug1228087 - Add go to copied link preference item

Although the copy text is “Go to copied URL” on the mocks here (https://bugzilla.mozilla.org/show_bug.cgi?id=1228087) I changed to “Go to Copied Link” so it’s consistent with the Today Widget text.

* Bug 1228087 - Move copiedURL logic to Pasteboard extension

copiedURL var will also be used to implement Go to copied URL ticket (https://bugzilla.mozilla.org/show_bug.cgi?id=1228087) it makes sense to share the same code between Today widget and main app.

* Bug 1228087 - Add description label to ButtonToast

To use the ButtonToast on this task it was necessary to add a description label to it.
Default values were added to make sure previous usage will keep working fine.

ToastDismissAfter struct item was removed since was not being used anywhere.

* Bug 1228087 - Display toast if there’s a copied link

https://bugzilla.mozilla.org/show_bug.cgi?id=1228087

ClipboardBarDisplayHandler was created to handle the logic if the bar should be displayed or not.

* Bug1228087 - Remove preference option

* Bug1228087 - Change clipboard display time

* Bug1228087 - Move strings to Strings.swift file

* Bug1228087 - Do not use UserDefaults to remember opened URL

There’s no need to save the last opened URL on the userDefault. To keep it simple it should be stored in memory

* Bug1228087 - Use already created method to display a toast

Clipboard toast was not using the correct method to display a toast.

* Bug1228087 - Move clipboardToast to ClipboardBarDisplayHandler

This was done to prevent BVC from getting to big.

* Bug1228087 - Make code more “swifty”

Better usage of guard, internal/external parameter names and declare the delegate as both used types.

* Fix SwiftLint errors