Bug 11339824 - ATTACH metadata (#2533)
* Part 1 - Add ability to attach a database to a BrowserDB instance

* Part 2 - Attach a DB called metadata.db to browser.db

If the DB doesn't exist it will be created

This cannot be done inside a transaction - this is a quirk of the `ATTACH` command, and because we reference the metadata table before we insert anything into it, through top sites and highlights, we cannot wait to create or attach the DB until we start to populate its tables, so we have to do it here.

Also, as `logins.db` also implements `BrowserDB`, we cannot do it during the init of `browser.db` unless we want to attach `logins.db` too. Perhaps a follow up should be filed to separate Logins and Browser

* Part 3 - Move all `page_metadata` table and index creation to create inside the attached DB.

We are keeping all of the old SQL around for migration purposes.

* Part 4 - update all tests to ensure they attach the metadata db before creating the `BrowserTable`

Otherwise everything goes wrong when it cannot find the attached DB to create metadata