Bug 1392317: calculated display bounds for text is too tight (#3075)
The row height of the footer is calculated from the estimated size of the text.
The calculated bounds of text are probably exact to the pixel edge; when
rendered in a label, the last letter can be too close to the edge and partially cut off
or ellipsis shown (depending on the uilabel setting).
Over-estimate the size needed for the label and this won't be a problem.