Bug 1370281 – Use async getters for UIPasteboard when using ClipboardBar. (#3067)
* Bug 1370281 — Add async getters to UIPasteboardExtensions.

Use Deferred<Maybe<T?>>; we have lots of tooling associated with Deferred<Maybe>>, so keep with that, though it does make it feel a little clunky when dealing with optionals.

* Bug 1370281 – Use async getters for UIPasteboard when using ClipboardBar.

This also tidies up the ClipboardBarDisplayHandler up a little, so as

 1. to reduce the number of times the paste board is fetched from
 2. not been too dependent on URL types (except in initial fetch of the URL)
 3. bail as quickly as possible if the clipboard bar is set as off in the settings.

This also moves to use absoluteDisplayString instead of the absolute string.