Bug 1394786: action extension UTI needs to be public.url (#3137)
BANDAID ONLY: this fixes everything except 1PW no longer shows as an option. A better solution to follow ASAP.

As of iOS11 we are getting this error reported when tapping extensions that
share the URL:
`Cannot load representation of type org.appextension.fill-browser-action`
There are 2 types of UTI we report: org.appextension.fill-browser-action and
public.url.
For some reason we returned the former in nearly all cases (not sure why, maybe
a legacy reason).
If itemForActivityType is changed to always return public.url, it works on iOS
10 and 11, and pw extensions still work ok.