Bug 1395547 — Clear remote-only tabs when disconnecting from FxA account. (#3199) r=rnewman, justindarc
* Bug 1395547 — Add FxADeviceId in SyncStateMachine.

Persist it in the scratchpad prefs.

This replaces the previous hack of poking deviceRegistration into a constructed prefs object.

* Bug 1395547 — Detect a change in our client record and reset the record if changed.

* Bug 1395547 — Only fill deferred once our tabs upload.

This ultimately did not fix the bug we were seeking to fix, though should now return correct completed stats.

* Bug 1395547 — Only remove the tabs from remote clients.

Currently we do not close tabs when disconnecting an account. This `clear` method now reflects that, and leaves rows in the tabs table that are open on this device.

* Bug 1395547 — Add a migration between storing the whole deviceRegistration to just storing the deviceID.

* Bug 1395547 — Addressed nits.