Bug 1433475 — Keyboard — enable up and down arrows in location bar. (#3706) r=farhan
This bug allows the user to select URL completions (from bookmarks and history) to be selected via the keyboard.

Desktop leaves the cursor in the location field, which means than horizontal cursor movement is possible.

We try to replicate that.

There do seem to be some difficulty supporting search suggestions, so this is rolled into another bug.