Bug 1475608 - Dark theme followup: home panels, status bar (#4044)
* Fix snackbar title. Not planning to theme snackbars (yet)

* Enable visible panels to change theme.
Also, ensure welcome screens within the panels do live theme updates.

* Change implicit unwrap Profile! in home panels.
The `profile` member var is required for these classes to function, it should be an init arg, not set on a consecutive line.
I am changing the init flow as part of theming, and am getting accidental breakage because my new init calls are triggering early-stage calls to `.profile` and the force-unwrap is fatal.

* More home panel theming.
- Move styling for classes into their applyThem(), then init() calls applyTheme() as does code that triggers a theme change
- Theme changes call reload data where needed
- TableView cell and header subclasses call applyTheme() in prepareForReuse().
Header/Footer subclasses only have a call to reset their title when table is reloaded, prepareForReuse() is the place to set common properties for all such cells when reused.

* left an ! in accidentally

* A few missed items:
- reading list missing a dark theme for text
- TwoLineTableViewCell wasn't using applyTheme() -cleanup item, just makes it easier to find color assignment code. I like it all in applyTheme() functions when possible.
- bookmarks back cell in dark theme wasn't using the right color