Merge Nimbledroid branch to master. (#4055)
* Added nimbledroid for perf testing

* PR #4055 numbledroid follow up. Use 'shared' for singleton access

* Temporarily omit specifying master branch only upload

Temporary change while we check for successful uploads

* Add Archive pre-action call to scheme-postbuild script (#4069)

Our test runner is trying to access a database for each test class that is stored in the `test-fixtures` directory of the  project. Our scheme defines a pre-action step for the Test action that copies the databases to the app bundle but it’s missing that step for the Archive action.

* Fix broken curl parameter
` {"code":400,"message":["`test_runner` was injected, please provide `test_identifiers`"]}`

* fix merge error

* Fix NDScenario startup not getting called

* Fix compile error post-merge

* Fix Profiler. to Profiler.shared?.

* remove bvc_did_appear due to problems