Bug 1496712 - TDM not selecting new tab, ensure completion blocks are called. (#4314)
The code works such that blocks are queued to be run when batchupdates is completed,
The code was setup so the completion blocks getting called is optional, and this bug is that the completion block for the new tab is not called, thus the the tab tray is not dismissed.
The fix is to ensure the pending completion blocks are always called.
There are too many state flags in this code for me to follow to see why a particular path is deciding not to call the completion blocks, but I am confident this step is required.