Bug 1499058 - TabDisplayManager use chained/imperative operation queue (#4323)
* Bug 1310071 - Download gifs properly

* enable tab reorder tests

* handle close all tabs, this should dismiss tabs tray and select a tab

* Fix fast deletion, the selected tab update is grabbing the wrong tab index, oops

* Fix deselecting a tab. Remove stateful dependancy on 'previous selected' tab.
The previous selected tab can be determined on-demand.

* Reformat `performChainedOperations()` to clarify recursion termination condition

* Add comments on how to use performChainedOperations

* Cleanup DragAndDropTests syntax and use localhost pages instead of remote

* Add non-UI tests to verify dragging is disabled when tab events arrive.

* Bug 1499058 - TabDisplayManager use chained/imperative operation queue
- remove all state management in TDM
- weakify all self in closures
- private mode toggling logic consolidated into TDM

* fix scrolling to selected tab

* Fixing a broken test for remove all tabs with toast.
TabManager had previously decoupled removal and selection, so I removed all had extra logic to work around
the selected tab changing during removal. This is not needed (and we should avoid having the tab manager do specific logic to
accomodate the tab display, the delegate functions should get params added that enable the display code to make decisions as to what to do).