Bug 1509383 - WKURLSchemeHander for all except reader mode (#4387)
* Refactor making the default configuration for a webview

* Custom URL scheme for all except Reader Mode

* fix non-XCUITests

* Use InternalSchemeHandler resource loading code to load css for error pages.

* Do not allow arbitrary internal resource loads, be explicit about the filename.

* Add InternalURL(URL) to encapsulate all internal scheme logic for URLs
rather than using URL extension. Extensions to built-in classes should
be more minimal that what was being done previously.
I think this makes searching the code for use of internal schemes
easier, makes refactoring this logic easier, and in most cases makes the
logic clearer to read (the author must organize their logic around
whether an url is internal or not, it is less 'mixed in' with
non-internal scheme logic).

* Remove downloadResource() handling for reader mode pages.

* Add UUID url param to authorize internal urls

* Workaround WKWebView restore error page redirection bugs.
WKWebView has a bug if the js changes the page location internally,
any attempt to redirect to the new URL needs to happen after a timeout.

* Fix page restore code. Only one case is missing,
an immediate show of an error page.
In this case, the sessionrestore.html needs to do
a `location.replace('internal://local/errorpage?url=...')`, rather than
a webview.load (or js `location=`) which will add a page to the history
stack.

* Fix restoring url that immediately triggers an error page.

* Fix tests

* make sure to authorize internal session resore urls

* address review request