Bug 1537709 - ContextMenuHelper retain cycle causing crash (#4638)
* Bug 1537709 - ContextMenuHelper retain cycle causing crash
To be clear, I can't reproduce the crash, but the crash report is reporting
a bad retain of the KVO observer, and there is an obvious bug in that code.
Fingers crossed this fixes it.

* Leak in TabContentScriptManager because it is not uninstalled from the webview

* On 12.1 sim there are crashes because this is required
` kvoInfo.observation?.invalidate() `
and also, the WKContentView is not the first item in the subviews array.