Fix #4952 - Low contrast menus when reduced transparency/inc. contrast are enabled (#5282)
* Night mode blur color, and reduce transparency

Previous color was a result of no defined background color
Blur effect in PhotonActionSheet disables on reduced transparency, enables otherwise
Light theme removes alpha on background to deal with above case

* Fixed title and subtitle text colors

Tint colors are modified by iOS when increased contrast is enabled — now, the behavior for PhotonActionSheetCell is accessible, in addition to being inline with ThemedWidgets.

* Night mode action foreground color -> white

This increases readability for when “increase contrast” is enabled.

* Disable theme bg alpha on reduced transprency

Also:
- flipped reduce transparency logic to be more readable -> (isEnabled) instead of (!isEnabled)
- Changed Theme ActionMenuColor to be inline with corresponding DarkTheme color, by just using defaultBackground