Fix #5476 - Support switching to opposite UA for iPad and iPhone (#5514)
* Move tab desktop sites code to own file,
and rename it ChangeUserAgent, as it now tracks whether
the UA is changed to the opposite UA.

* Calculating the UA a runtime using a hidden webview no longer works as of iOS13.
On iPad it returns a mobile UA, not desktop UA. As a result,
hard-code the user agent strings.

* Remove desktopSite from TabState, it appears to be unused.

* Rename Tab.desktopSite to changedUserAgent and change related code.
Change logic to use `oppositeUserAgent()` for changing user agent.

* Remove complicated reload() logic. We need to simplify as much code as possible for moving to proper iOS 13 desktop site mode support.
This supported forcing a reload ignoring the cache, but in my testing a regular reload works fine with a changed UA.

* Update the logic to try remove `m.` from the URL to do this only on iPhone.
This is for the case where the site is an `m.` url and the user switches to desktop site,
without changing the url it won't load the desktop site.