Fix #6376: migration of pre-rust bookmarks breaking tests (#6377)
For migration of users to Rust bookmarks backend, if users are signed in to sync then next sync will download their bookmarks.
For users not using sync, they will lose their bookmarks, this code extracts their bookmarks from browser.db into places.db.
This code is old at this point, we have no test case for it, and it has to run at a very specifc point in app init, which is hard to ensure given the async startup nature of the new firefox account system.
I think the risk of keeping this outweighs the risk of removing it.