Fix #6390 and #6376: restore bookmark migration code to support tests (#6393)
We have tests using browser.db for their bookmarks, which need to be
migrated to places.db for them to run. This means we have a test case
for the migration code, which I didn't realize, and not having a test case
for the migration code was a rationale I used to remove it.