Copy appservices local dev resources, rather than symlinking. (#8081)
Now that we're using `carthage copy-framework` to bundle the Application
Services framework, it doesn't seem possible to use a symlink for local
development - carthage tries to turn it into a relative path and gets
mightily confused.

As an alternative, this modifies the local-dev scripts to rsync across
the build artifacts into `./Carthage/Build/`. It's less convenient than
before, but at least it works.