Fixes #8614 - Adds a device-type label to the URL (#8615)
* Issue #1905 - Reorganizes the code comments.

The code comments were organized for the SUMO URL,
but they didn't take into account the modifications
which have been added for webcompat. We add the relevant
comments.

* Issue #1905 - Makes code formatting consistent

* Issue #1905 - Adds identification of the device type

The reporting URL will now send a device-type extra label.
When the device is an iPad, it will become easier to classify
for webcompat folks as the iPad usually sends a desktop UA.