FXIOS-1771 inactive tabs in background (#8903)
Implemented less intrusive background mode for inactive tabs.

User has X amount of tabs opened out of which some are opened on same day, some are more than 4 days old, and rest are more than 30 days old

Types of Tabs: Normal, Inactive and Recently Closed

When the feature first gets enabled we start marking the tabs that would become inactive or needs closing. Marking of tabs happens when:

    - User interacts with a tab after feature has enabled
    - User opens tab tray
    - User adds / removes a tab

If none of this happens after the feature first gets enabled then we mark the tabs for closing on second launch. Reason for this is to avoid automatic closing of tabs while user in their session and since user never interacted with tabs in their last session we treat this session as a first time feature enable.