FXIOS-2981 - Business Logic and UI for Search Term Groups [iPhone and iPad]  (#9025)
Added Search Term Groups in Tabs Tray which has the following sections and is only applicable for Normal / Active tabs and not for private mode.

Note: This PR also turns on search term metadata collection (FXIOS-3063)

## a) Search Term Groups
This section contains grouped tabs formed by the search term entered by the user

iPhone: Number of items shown on screen in a group are 2 unless we go into landscape mode. 
iPad: Number of items shown on screen in a group are 4

Min number of items that can make a group for both iPhone and iPad are 2. If only a single Tab is part of search metadata group then we don't do anything with it and simply make it a part of Regular Tab

Usage:
-- User opens a new tab and searches for a _Search Term 1_
-- This _Search Term 1_ gets recorded as tab history metadata 
-- User opens another tab and searches for a _Search Term 1_ (same search) but now taps on a item on search page. This is a search follow-on scenario
-- We combine both tabs into **Search Term 1 Group** in Tab Tray
-- Search Groups are collection of tabs when user searches for a particular item from *Awesome Bar* 

## b) Regular Tabs

These are normal tabs that are not part of a group and user simply opens them by entering URL or long pressing on a link. 

## c) Inactive Tabs

Inactive tabs (if enabled) should put a tab to inactive mode after 4 days. The number of days after which a tab gets inactive might change in future. 

**Whats missing:** 
Telemetry and UI polish