#9576 ⁃ [Refactor] [v96] - search ad count (#9634)
This is an improvement to our ad search count.

It no longer relies on redirects when the user taps on an ad and rather checks the last host to compare with all other redirects. This allows us to not play catch with capturing redirect URLs that are associated with a particular search engine.

Co-authored-by: lmarceau <lmarceau@mozilla.com>