Bugfix #9666 #9477 [v97] Login search previous query fix (#9714)
When multiple searches are created, if the previous query is cancelled with an empty success result, this causes the UI to stop displaying a loading screen. This is very confusing to users, as it makes them believe the search is not working properly. This PR changes that to do not fulfill previous query with a success when replacing password search with a new term. Fixes #9666