Refactor FXIOS-7699 [v122] Continue migration by using UIButton.Configuration in PrimaryRoundedButton (#17616)
* Renamed ResizableButton to LegacyResizableButton

* Added a new ResizableButton class using UIButton.Configuration

* Added new helper function to set a button's font through configuration

The way to set the font using UIButton.Configuration is pretty verbose so this helper function will help reduce code duplication once migrating the rest of the codebase

* Updated PrimaryRoundedButton to use the new ResizableButton class

* Updated PrimaryRoundedButton usage in FirefoxAccountSignInViewController

Using `withRenderingMode(.alwaysTemplate)` on the button's image so it gets tinted like the button's title when highlighted, interpolating from the base colour.
(This fixes a bug where the theme was not being properly taken into account in the signin page. If the button was highlighted when using dark mode, the image was white instead of dark)

* Updated usage of PrimaryRoundedButton in DefaultBrowsingOnboardingViewController

Added an extension `NSDirectionalEdgeInsets.init(equalInset:)` based on the existing one for `UIEdgeInsets`

* Auto stash before merge of "feature/migrate-primarybutton-to-uibutton-configuration" and "upstream/main"

* Revert accidental changes commited to Package.resolved

* Remove unnecessary NSDirectionalEdgeInsets extension

* Match update of configuration with LinkButton for concistency

* Use textInverted for all states of PrimaryRoundedButton

* Cleanup setting updated configuration

* Added a comment to tell that extension is only a temporary fix