[focus-ios] Fix missing buttons in Find In Page bar (https://github.com/mozilla-mobile/focus-ios/pull/3894)
* Fix invisible Find Previous & Find Next buttons

Move the corresponding icon assets to the DesginSystem module so that
they can be found at runtime.

Fixes https://github.com/mozilla-mobile/focus-ios/issues/3699

* Make Find Previous / Find Next buttons initially disabled

When the user hasn't entered any text, the previous/next buttons were
not made disabled, which could be a little bit misleading and was not
consistent with the case where there are no search results.

---------

Co-authored-by: isabelrios <isabelrios@gmail.com>