Refactor FXIOS-9246 - Qualify MozillaAppServices imports (#20479)
This helps reduce the chance of name collisions caused by
`MozillaAppServices` defining types with the same names as
Firefox for iOS types.

Because of how we package Application Services, it can be
difficult to tell where each type comes from, especially
when other modules re-export those imports. This commit
aims to help with that as well.