Build - Run bin/nimbus-fml.sh with a clean environment (#21385)
Run Script build phases inherit all Xcode build settings as
environment variables, including the compiler, linker, and flags,
for the target architecture.

This means that, e.g., when building Firefox or Focus for
an iOS device, Xcode will set build environment variables for
the iOS target; when building Firefox for the Simulator,
Xcode will set build variables for the Simulator target; etc.

Unfortunately, this is the wrong thing to do when invoking the
`nimbus-fml` script with `MOZ_APPSERVICES_LOCAL` set. In that case,
the script will invoke `cargo run`, which will see the Xcode-set
build variables, and try to cross-compile and run the local
`nimbus-fml` for the target architecture (iOS, Simulator, etc.).

Put another way, Cargo is trying (and failing) to build and run
an iOS binary on a Mac.

This commit fixes `bin/nimbus-fml.sh` to always run with a
clean environment, except for the variables that it needs.