[MOB-3334] Update Ecosia domain and corresponding UA
Having an Ecosia dependency within BrowserKit is not recommended as we want BrowserKit (and our Ecosia framework) as isolated as possible. Hardcoding is never a good idea but we do have other hardcoded strings in the UserAgent.swift class so it is worth the change in favor of a less import.